upload song

+++ backend/src/controllers/History.controller.js
@@ -2,7 +2,12 @@ const historymodel = require('../models/History.model');
 
 const getallhistory = async(req,res)=>{
     try{
-        const history =await historymodel.find().populate('songId'); // ye populate kya karega ki songId ke andar jo bhi data hoga wo aa jayega
+        const history =await historymodel.find({
+            userId:req.user.id,
+            
+        }).populate('songId');
+            
+         // ye populate kya karega ki songId ke andar jo bhi data hoga wo aa jayega
         const validhistory = history.filter(history => history.songId!==null); // ye filter method se hum valid history ko filter karenge taki hume sirf valid history hi mile
         res.status(200).json(validhistory);
     }catch(err){
@@ -12,7 +17,11 @@ const getallhistory = async(req,res)=>{
 
 const addhistory = async(req,res)=>{
     try{
-        const history = await historymodel.create(req.body);
+        const history = await historymodel.create({
+            userId:req.user.id,
+            songId:req.body.songId,
+            
+        });
         res.status(201).json(history);
     }catch(err){
         res.status(500).json({message: err.message});
@@ -22,7 +31,11 @@ const addhistory = async(req,res)=>{
 
 const deletehistory = async(req,res)=>{
     try{
-        const history = await historymodel.findByIdAndDelete(req.params.id);
+        const history = await historymodel.findOneAndDelete({
+            _id:req.params.id,
+            userId:req.user.id
+            
+        });
         if(!history){
             return res.status(404).json({message: "History not found"});
         }
@@ -34,7 +47,7 @@ const deletehistory = async(req,res)=>{
 
 const deleteAllhistory = async(req,res)=>{
     try{
-        await historymodel.deleteMany({}); // ye deleteMany method se hum saari history ko delete kar denge 
+        await historymodel.deleteMany({userId:req.user.id}); // ye deleteMany method se hum saari history ko delete kar denge 
         res.status(200).json({message: "All history deleted successfully"});
     }catch(error){
         res.status(500).json({message: error.message});
